+++ internal/mesh/internal/types/http_route_test.go
@@ -4,14 +4,10 @@
 package types
 
 import (
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/durationpb"
-	"google.golang.org/protobuf/types/known/wrapperspb"
 
 	"github.com/hashicorp/consul/internal/resource"
 	"github.com/hashicorp/consul/internal/resource/resourcetest"
@@ -35,6 +31,7 @@ func TestMutateHTTPRoute(t *testing.T) {
 			WithTenancy(tc.routeTenancy).
 			WithData(t, tc.route).
 			Build()
+		resource.DefaultIDTenancy(res.Id, nil, resource.DefaultNamespacedTenancy())
 
 		err := MutateHTTPRoute(res)
 
@@ -200,14 +197,17 @@ func TestMutateHTTPRoute(t *testing.T) {
 
 func TestValidateHTTPRoute(t *testing.T) {
 	type testcase struct {
-		route     *pbmesh.HTTPRoute
-		expectErr string
+		routeTenancy *pbresource.Tenancy
+		route        *pbmesh.HTTPRoute
+		expectErr    string
 	}
 
 	run := func(t *testing.T, tc testcase) {
 		res := resourcetest.Resource(pbmesh.HTTPRouteType, "api").
+			WithTenancy(tc.routeTenancy).
 			WithData(t, tc.route).
 			Build()
+		resource.DefaultIDTenancy(res.Id, nil, resource.DefaultNamespacedTenancy())
 
 		// Ensure things are properly mutated and updated in the inputs.
 		err := MutateHTTPRoute(res)
@@ -844,6 +844,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 	// Add common parent refs test cases.
 	for name, parentTC := range getXRouteParentRefTestCases() {
 		cases["parent-ref: "+name] = testcase{
+			routeTenancy: parentTC.routeTenancy,
 			route: &pbmesh.HTTPRoute{
 				ParentRefs: parentTC.refs,
 			},
@@ -859,6 +860,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 			})
 		}
 		cases["backend-ref: "+name] = testcase{
+			routeTenancy: backendTC.routeTenancy,
 			route: &pbmesh.HTTPRoute{
 				ParentRefs: []*pbmesh.ParentReference{
 					newParentRef(pbcatalog.ServiceType, "web", ""),
@@ -875,319 +877,6 @@ func TestValidateHTTPRoute(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			run(t, tc)
 		})
-	}
-}
-
-type xRouteParentRefMutateTestcase struct {
-	routeTenancy *pbresource.Tenancy
-	refs         []*pbmesh.ParentReference
-	expect       []*pbmesh.ParentReference
-}
-
-func getXRouteParentRefMutateTestCases() map[string]xRouteParentRefMutateTestcase {
-	newRef := func(typ *pbresource.Type, tenancyStr, name string) *pbresource.Reference {
-		return resourcetest.Resource(typ, name).
-			WithTenancy(newTestTenancy(tenancyStr)).
-			Reference("")
-	}
-
-	newParentRef := func(typ *pbresource.Type, tenancyStr, name, port string) *pbmesh.ParentReference {
-		return &pbmesh.ParentReference{
-			Ref:  newRef(typ, tenancyStr, name),
-			Port: port,
-		}
-	}
-
-	return map[string]xRouteParentRefMutateTestcase{
-		"parent ref tenancies defaulted": {
-			routeTenancy: newTestTenancy("foo.bar"),
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "", "api", ""),
-				newParentRef(pbcatalog.ServiceType, ".zim", "api", ""),
-				newParentRef(pbcatalog.ServiceType, "gir.zim", "api", ""),
-			},
-			expect: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "foo.bar", "api", ""),
-				newParentRef(pbcatalog.ServiceType, "foo.zim", "api", ""),
-				newParentRef(pbcatalog.ServiceType, "gir.zim", "api", ""),
-			},
-		},
-	}
-}
-
-type xRouteBackendRefMutateTestcase struct {
-	routeTenancy *pbresource.Tenancy
-	refs         []*pbmesh.BackendReference
-	expect       []*pbmesh.BackendReference
-}
-
-func getXRouteBackendRefMutateTestCases() map[string]xRouteBackendRefMutateTestcase {
-	newRef := func(typ *pbresource.Type, tenancyStr, name string) *pbresource.Reference {
-		return resourcetest.Resource(typ, name).
-			WithTenancy(newTestTenancy(tenancyStr)).
-			Reference("")
-	}
-
-	newBackendRef := func(typ *pbresource.Type, tenancyStr, name, port string) *pbmesh.BackendReference {
-		return &pbmesh.BackendReference{
-			Ref:  newRef(typ, tenancyStr, name),
-			Port: port,
-		}
-	}
-
-	return map[string]xRouteBackendRefMutateTestcase{
-		"backend ref tenancies defaulted": {
-			routeTenancy: newTestTenancy("foo.bar"),
-			refs: []*pbmesh.BackendReference{
-				newBackendRef(pbcatalog.ServiceType, "", "api", ""),
-				newBackendRef(pbcatalog.ServiceType, ".zim", "api", ""),
-				newBackendRef(pbcatalog.ServiceType, "gir.zim", "api", ""),
-			},
-			expect: []*pbmesh.BackendReference{
-				newBackendRef(pbcatalog.ServiceType, "foo.bar", "api", ""),
-				newBackendRef(pbcatalog.ServiceType, "foo.zim", "api", ""),
-				newBackendRef(pbcatalog.ServiceType, "gir.zim", "api", ""),
-			},
-		},
-	}
-}
-
-type xRouteParentRefTestcase struct {
-	refs      []*pbmesh.ParentReference
-	expectErr string
-}
-
-func getXRouteParentRefTestCases() map[string]xRouteParentRefTestcase {
-	return map[string]xRouteParentRefTestcase{
-		"no parent refs": {
-			expectErr: `invalid "parent_refs" field: cannot be empty`,
-		},
-		"parent ref with nil ref": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", ""),
-				{
-					Ref:  nil,
-					Port: "http",
-				},
-			},
-			expectErr: `invalid element at index 1 of list "parent_refs": invalid "ref" field: missing required field`,
-		},
-		"parent ref with bad type ref": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", ""),
-				newParentRef(pbcatalog.WorkloadType, "api", ""),
-			},
-			expectErr: `invalid element at index 1 of list "parent_refs": invalid "ref" field: invalid "type" field: reference must have type catalog.v2beta1.Service`,
-		},
-		"parent ref with section": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", ""),
-				{
-					Ref:  resourcetest.Resource(pbcatalog.ServiceType, "web").Reference("section2"),
-					Port: "http",
-				},
-			},
-			expectErr: `invalid element at index 1 of list "parent_refs": invalid "ref" field: invalid "section" field: section cannot be set here`,
-		},
-		"duplicate exact parents": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", "http"),
-				newParentRef(pbcatalog.ServiceType, "api", "http"),
-			},
-			expectErr: `invalid element at index 1 of list "parent_refs": invalid "port" field: parent ref "catalog.v2beta1.Service/default.local.default/api" for port "http" exists twice`,
-		},
-		"duplicate wild parents": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", ""),
-				newParentRef(pbcatalog.ServiceType, "api", ""),
-			},
-			expectErr: `invalid element at index 1 of list "parent_refs": invalid "port" field: parent ref "catalog.v2beta1.Service/default.local.default/api" for wildcard port exists twice`,
-		},
-		"duplicate parents via exact+wild overlap": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", "http"),
-				newParentRef(pbcatalog.ServiceType, "api", ""),
-			},
-			expectErr: `invalid element at index 1 of list "parent_refs": invalid "port" field: parent ref "catalog.v2beta1.Service/default.local.default/api" for ports [http] covered by wildcard port already`,
-		},
-		"duplicate parents via exact+wild overlap (reversed)": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", ""),
-				newParentRef(pbcatalog.ServiceType, "api", "http"),
-			},
-			expectErr: `invalid element at index 1 of list "parent_refs": invalid "port" field: parent ref "catalog.v2beta1.Service/default.local.default/api" for port "http" covered by wildcard port already`,
-		},
-		"good single parent ref": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", "http"),
-			},
-		},
-		"good muliple parent refs": {
-			refs: []*pbmesh.ParentReference{
-				newParentRef(pbcatalog.ServiceType, "api", "http"),
-				newParentRef(pbcatalog.ServiceType, "web", ""),
-			},
-		},
-	}
-}
-
-type xRouteBackendRefTestcase struct {
-	refs      []*pbmesh.BackendReference
-	expectErr string
-}
-
-func getXRouteBackendRefTestCases() map[string]xRouteBackendRefTestcase {
-	return map[string]xRouteBackendRefTestcase{
-		"no backend refs": {
-			expectErr: `invalid "backend_refs" field: cannot be empty`,
-		},
-		"backend ref with nil ref": {
-			refs: []*pbmesh.BackendReference{
-				newBackendRef(pbcatalog.ServiceType, "api", ""),
-				{
-					Ref:  nil,
-					Port: "http",
-				},
-			},
-			expectErr: `invalid element at index 0 of list "rules": invalid element at index 1 of list "backend_refs": invalid "backend_ref" field: invalid "ref" field: missing required field`,
-		},
-		"backend ref with bad type ref": {
-			refs: []*pbmesh.BackendReference{
-				newBackendRef(pbcatalog.ServiceType, "api", ""),
-				newBackendRef(pbcatalog.WorkloadType, "api", ""),
-			},
-			expectErr: `invalid element at index 0 of list "rules": invalid element at index 1 of list "backend_refs": invalid "backend_ref" field: invalid "ref" field: invalid "type" field: reference must have type catalog.v2beta1.Service`,
-		},
-		"backend ref with section": {
-			refs: []*pbmesh.BackendReference{
-				newBackendRef(pbcatalog.ServiceType, "api", ""),
-				{
-					Ref:  resourcetest.Resource(pbcatalog.ServiceType, "web").Reference("section2"),
-					Port: "http",
-				},
-			},
-			expectErr: `invalid element at index 0 of list "rules": invalid element at index 1 of list "backend_refs": invalid "backend_ref" field: invalid "ref" field: invalid "section" field: section cannot be set here`,
-		},
-		"backend ref with datacenter": {
-			refs: []*pbmesh.BackendReference{
-				newBackendRef(pbcatalog.ServiceType, "api", ""),
-				{
-					Ref:        newRef(pbcatalog.ServiceType, "db"),
-					Port:       "http",
-					Datacenter: "dc2",
-				},
-			},
-			expectErr: `invalid element at index 0 of list "rules": invalid element at index 1 of list "backend_refs": invalid "backend_ref" field: invalid "datacenter" field: datacenter is not yet supported on backend refs`,
-		},
-		"good backend ref": {
-			refs: []*pbmesh.BackendReference{
-				newBackendRef(pbcatalog.ServiceType, "api", ""),
-				{
-					Ref:  newRef(pbcatalog.ServiceType, "db"),
-					Port: "http",
-				},
-			},
-		},
-	}
-}
-
-type xRouteTimeoutsTestcase struct {
-	timeouts  *pbmesh.HTTPRouteTimeouts
-	expectErr string
-}
-
-func getXRouteTimeoutsTestCases() map[string]xRouteTimeoutsTestcase {
-	return map[string]xRouteTimeoutsTestcase{
-		"bad request": {
-			timeouts: &pbmesh.HTTPRouteTimeouts{
-				Request: durationpb.New(-1 * time.Second),
-			},
-			expectErr: `invalid element at index 0 of list "rules": invalid "timeouts" field: invalid "request" field: timeout cannot be negative: -1s`,
-		},
-		"bad idle": {
-			timeouts: &pbmesh.HTTPRouteTimeouts{
-				Idle: durationpb.New(-1 * time.Second),
-			},
-			expectErr: `invalid element at index 0 of list "rules": invalid "timeouts" field: invalid "idle" field: timeout cannot be negative: -1s`,
-		},
-		"good all": {
-			timeouts: &pbmesh.HTTPRouteTimeouts{
-				Request: durationpb.New(1 * time.Second),
-				Idle:    durationpb.New(3 * time.Second),
-			},
-		},
-	}
-}
-
-type xRouteRetriesTestcase struct {
-	retries   *pbmesh.HTTPRouteRetries
-	expectErr string
-}
-
-func getXRouteRetriesTestCases() map[string]xRouteRetriesTestcase {
-	return map[string]xRouteRetriesTestcase{
-		"bad conditions": {
-			retries: &pbmesh.HTTPRouteRetries{
-				OnConditions: []string{"garbage"},
-			},
-			expectErr: `invalid element at index 0 of list "rules": invalid "retries" field: invalid element at index 0 of list "on_conditions": not a valid retry condition: "garbage"`,
-		},
-		"good all": {
-			retries: &pbmesh.HTTPRouteRetries{
-				Number:       wrapperspb.UInt32(5),
-				OnConditions: []string{"internal"},
-			},
-		},
-	}
-}
-
-func newRef(typ *pbresource.Type, name string) *pbresource.Reference {
-	return newRefWithTenancy(typ, nil, name)
-}
-
-func newRefWithTenancy(typ *pbresource.Type, tenancy *pbresource.Tenancy, name string) *pbresource.Reference {
-	if tenancy == nil {
-		tenancy = resource.DefaultNamespacedTenancy()
-	}
-	return resourcetest.Resource(typ, name).
-		WithTenancy(tenancy).
-		Reference("")
-}
-
-func newBackendRef(typ *pbresource.Type, name, port string) *pbmesh.BackendReference {
-	return &pbmesh.BackendReference{
-		Ref:  newRef(typ, name),
-		Port: port,
-	}
-}
-
-func newParentRef(typ *pbresource.Type, name, port string) *pbmesh.ParentReference {
-	return newParentRefWithTenancy(typ, nil, name, port)
-}
-
-func newParentRefWithTenancy(typ *pbresource.Type, tenancy *pbresource.Tenancy, name, port string) *pbmesh.ParentReference {
-	return &pbmesh.ParentReference{
-		Ref:  newRefWithTenancy(typ, tenancy, name),
-		Port: port,
-	}
-}
-
-func newTestTenancy(s string) *pbresource.Tenancy {
-	parts := strings.Split(s, ".")
-	switch len(parts) {
-	case 0:
-		return resource.DefaultClusteredTenancy()
-	case 1:
-		v := resource.DefaultPartitionedTenancy()
-		v.Partition = parts[0]
-		return v
-	case 2:
-		v := resource.DefaultNamespacedTenancy()
-		v.Partition = parts[0]
-		v.Namespace = parts[1]
-		return v
-	default:
-		return &pbresource.Tenancy{Partition: "BAD", Namespace: "BAD", PeerName: "BAD"}
 	}
 }
