+++ internal/mesh/internal/types/tcp_route.go
@@ -64,7 +64,7 @@ func ValidateTCPRoute(res *pbresource.Resource) error {
 
 	var merr error
 
-	if err := validateParentRefs(route.ParentRefs); err != nil {
+	if err := validateParentRefs(res.Id, route.ParentRefs); err != nil {
 		merr = multierror.Append(merr, err)
 	}
